pyproject.toml
[project]
name = "seedling"
version = "0.2.0"
description = "A simple library for intent recognition using LLMs."
authors = [{name = "Tahmid Efaz"}]
license = {file = "LICENSE"}

dependencies = [
    "PyYAML==6.0.1",
    "jsonschema==4.21.1",
    "openai==1.23.6",
]


[project.optional-dependencies]
dev = [
    "flake8",
    "pylint",
]
